requirements.txt
streamlit==1.34.0
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==1.4.46
scikit-learn==1.1.3
nltk==3.8.1
python-dotenv==1.0.1
bcrypt==4.1.2
pyjwt==2.8.0
pandas<2.0.0
beautifulsoup4==4.12.3
requests==2.31.0
pytest==7.4.4
torch==2.0.1
transformers==4.41.0
sentence-transformers==4.0.2
scipy>=1.7.0
rasa==3.6.16
rasa-sdk==3.6.2
psycopg2-binary==2.9.9
alembic==1.13.1 